Extract superclass for command runners
